requirements.txt
# requirements.txt


fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.0
chromadb==0.4.24
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
python-multipart==0.0.6